Updated members.js initialization
refs https://github.com/TryGhost/membersjs/issues/6

Previously the members.js script was only being loaded for signed in members, this updates the script initialization to always load on theme and initialize it with relevant site and members data which needs to be used by script to manage its UI state and interactions.

The data being passed through to script is not complete and needs discussion on alternate or better ways this can be passed-in to avoid adding extra code other than members.js script if possible. Also additional data like member subscriptions might need to be introduced as well here later.